pyproject.toml
[tool.poetry]
name = "my-project"
version = "0.1.0"
description = ""
authors = ["Alex"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.5"
psycopg2-binary = "^2.9.1"
django-crispy-forms = "^1.12.0"
django-allauth = "^0.44.0"
Pillow = "^8.3.0"
stripe = "^2.58.0"
django-debug-toolbar = "^3.2.1"
whitenoise = "^5.2.0"
gunicorn = "^20.1.0"
dj-database-url = "^0.5.0"
djangorestframework = "^3.12.4"
django-environ = "^0.4.5"
PyYAML = "^5.4.1"
uritemplate = "^3.0.1"
django-extensions = "^3.1.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.12.1"
pre-commit = "^2.13.0"
flake8 = "^3.9.2"
mypy = "^0.910"
isort = "^5.9.1"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 100

[tool.black]
line-length = 100
target-version = ['py37']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | staticfiles
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
